Code refactor

+++ src/main/scala/Algo/Convolution.scala
@@ -1,11 +1,11 @@
 import breeze.linalg.DenseMatrix
 
-class Convolution (imageMatrix :DenseMatrix[Int], convolutionKernel :DenseMatrix[Double]) {
+class Convolution (imageMatrix :DenseMatrix[Double], convolutionKernel :DenseMatrix[Double]) {
 
-    def run () :DenseMatrix[Int]  = {
+    def run () :DenseMatrix[Double]  = {
         assert(convolutionKernel.cols == 3)
         assert(convolutionKernel.rows == 3)
-        val outMatrix =  DenseMatrix.zeros[Int](imageMatrix.rows , imageMatrix.cols)
+        val outMatrix =  DenseMatrix.zeros[Double](imageMatrix.rows , imageMatrix.cols)
         for {
             i <- 1 until outMatrix.rows -1
             j <- 1 until outMatrix.cols -1
@@ -14,7 +14,7 @@ class Convolution (imageMatrix :DenseMatrix[Int], convolutionKernel :DenseMatrix
         adjustImageOut(outMatrix)
     }
 
-    private def adjustImageOut(matrix: DenseMatrix[Int]) : DenseMatrix[Int] =
+    private def adjustImageOut(matrix: DenseMatrix[Double]) : DenseMatrix[Double] =
         matrix.map( elem => if ( elem > 255 ) 255 else
                             if ( elem < 0 ) 0 else elem)
 }
